services/docx-redliner/requirements.txt
python-docx>=1.1.0
diff-match-patch>=20230430
openai>=1.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0

